+++ CHANGELOG.md
@@ -45,6 +45,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 #### :white_check_mark: Validation
 #### :bug: Bugfixes
 * Fix selection of best background source when starting on a zoomed-out view ([#9325])
+* Fix leaking of localized strings in combo fields when taginfo service is unavailable ([#9342])
 #### :rocket: Presets
 * Support tagging schema v5 ([#9320]):
   * Add new field type `colour` ([schema-builder#38], [#8782])
@@ -57,6 +58,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 [#9294]: https://github.com/openstreetmap/iD/issues/9294
 [#9320]: https://github.com/openstreetmap/iD/pull/9320
 [#9325]: https://github.com/openstreetmap/iD/issues/9325
+[#9342]: https://github.com/openstreetmap/iD/issues/9342
 [schema-builder#38]: https://github.com/ideditor/schema-builder/pull/38
 
 # 2.22.0